generate gemm id for each type
